requirements.txt
Pillow>=10.0.0
requests>=2.28.0
pyinstaller>=6.0.0
PySide6>=6.5.0
